use autossh instead of ssh
This change adjusts the script to use autossh instead of simply ssh. By
using autossh we presumably get better availability as it is able to
reestablish broken sessions. Sessions can be interrupted for various
reasons -- some of which completely out of our control -- and so playing
safe seems to be the sensible approach.