generalize script by making host configurable
Until now the script had a hard coded host. That constraint limits its
flexibility tremendously.
This change makes the host configurable by means of the SSH_HOST
variable. As typical for OpenRC scripts, this variable should be
configured in a /etc/conf.d/ file of the same name as the service.